fix: cast wifi_freq to float instead of int in wifi_frequency_adapter

+++ adapters/wifi_frequency_adapter.py
@@ -37,5 +37,5 @@ class WifiFrequencyAdapter(MeasurementAdapter):
             }
 
         return {
-            "wifi_freq": int(freq_match.group(1)),
+            "wifi_freq": float(freq_match.group(1)),
         }
